requirements.txt
python-dotenv
aiohttp
feedparser
beautifulsoup4
google-generativeai
python-telegram-bot
aiosqlite
pytz
backoff
tweepy
Pillow
structlog
prometheus-client
psutil
deep-translator
# AI / NLP Dependencies
khmer-nltk
# nudenet  # Commented out to avoid massive download/build issues in this env, fallback is implemented
# transformers # Commented out to avoid massive download/build issues in this env, fallback is implemented
# torch